Baremetal FVP RDN2 - Wrong value of the Memory access properties (#353)
This change fixes the test 840 failure with FVP RDN2 baremetal testing. 

SBSA level 3 rule PCI_IC_11 requires root complexes to support IO coherency and to be in the same shareability domain as the CPUs.  Therefore, the current value of the RC Memory access properties (0x0 - not coherent) is wrong, it should be changed to 0x1 (fully coherent)